back-end/pyproject.toml
[tool.poetry]
name = "automicro-bot"
version = "0.1.0"
description = "AutoMicro-Bot — Floating AI Chatbot Backend"
authors = ["AutoMicro <automicro@local>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { version = "^0.32.0", extras = ["standard"] }
langchain = "^0.3.0"
langchain-ollama = "^0.2.0"
langgraph = "^0.2.0"
aiosqlite = "^0.20.0"
python-dotenv = "^1.0.0"
pydantic = "^2.9.0"
websockets = "^13.0"
langchain-openai = ">=0.3.0"
langgraph-checkpoint-sqlite = "<2.0.0"
langchain-anthropic = "<1.0.0"
chromadb = "^1.5.5"
sentence-transformers = "^5.3.0"

[tool.poetry.dev-dependencies]
httpx = "^0.27.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
start = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"
